kilosort4 updates
Added kilosort4 to the environment list rather than using the docker. This allows for more customization and it's easy to use the GUI. Added Kilo3_Tetrode_Example.ipynb to test integration. Users need to install cuda drivers and compatible torch distribution based on their machines. The test script is more of a playground.